git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@23977 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/clients/http/FirstTimeWizardToadlet.java
@@ -188,7 +188,7 @@ public class FirstTimeWizardToadlet extends Toadlet {
 			
 			Option sizeOption = config.get("node").getOption("outputBandwidthLimit");
 			if(!sizeOption.isDefault()) {
-				long current = (Integer)sizeOption.getValue();
+				int current = (Integer)sizeOption.getValue();
 				result.addChild("option", new String[] { "value", "selected" }, new String[] { SizeUtil.formatSize(current), "on" }, l10n("currentSpeed")+" "+SizeUtil.formatSize(current)+"/s");
 			}
